requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
email-validator
bcrypt
alembic
gunicorn
python-jose
